metric detail support dubbo sla method Name

+++ hera-all/hera-monitor/hera-monitor-service/src/main/java/com/xiaomi/mone/monitor/service/prometheus/PrometheusService.java
@@ -343,7 +343,7 @@ public class PrometheusService {
         if (!CollectionUtils.isEmpty(result)) {
             for (MetricDataSetVector metricDataVector : result) {
                 Metric metric = metricDataVector.getMetric();
-                if (Double.valueOf(metricDataVector.getValue().get(1)).intValue() == 0) {
+                if (Double.valueOf(metricDataVector.getValue().get(1)) == 0d) {
                     continue;
                 }
